requirements.txt
google-generativeai
gspread
oauth2client
pandas
requests
pygooglenews
python-dotenv
feedparser
streamlit
beautifulsoup4